Modified documentation, added LineSegment class
Added some HTML tidbits to the maths package for some fancier
documentation.
Removed the existing mandelbrot method, renamed the julia method to the
mandelbrot method, and extended it's documentation to explain how to use
it to calculate the mandelbrot set.
Removed power method, as it is redundant compared to the Math.pow()
default method.